Override the OWL RL class to use a named graph
