+++ octadocs/conversions.py
@@ -1,9 +1,10 @@
-from typing import Optional
+from typing import Optional, Iterator
 
 import rdflib
 from mkdocs.structure.pages import Page
-from octadocs.environment import src_path_to_iri
+from octadocs.types import Triple, Quad, LOCAL
 from octadocs.query import SelectResult, query
+from rdflib import URIRef
 
 
 def iri_by_page(page: Page) -> rdflib.URIRef:
@@ -30,3 +31,19 @@ def get_page_title_by_iri(
         return rows[0]['title']
 
     return None
+
+
+def triples_to_quads(
+    triples: Iterator[Triple],
+    graph: rdflib.URIRef,
+) -> Iterator[Quad]:
+    """Convert sequence of triples to sequence of quads."""
+    yield from (
+        triple.as_quad(graph)
+        for triple in triples
+    )
+
+
+def src_path_to_iri(src_path: str) -> URIRef:
+    """Convert src_path of a file to a Zet IRI."""
+    return URIRef(f'{LOCAL}{src_path}')
