+++ octadocs/macros.py
@@ -8,8 +8,8 @@ from unittest.mock import patch
 import pydotplus
 import rdflib
 from mkdocs_macros.plugin import MacrosPlugin
-from octadocs.conversions import iri_by_page
-from octadocs.environment import iri_to_url, src_path_to_iri
+from octadocs.conversions import iri_by_page, src_path_to_iri
+from octadocs.environment import iri_to_url
 from octadocs.query import query
 from rdflib.plugins.sparql.processor import SPARQLResult
 from rdflib.tools.rdf2dot import rdf2dot
