+++ octadocs/octiron/plugins/base.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import Iterator, Optional
 
 import rdflib
-from octadocs.octiron.types import Context, Triple
+from octadocs.types import Context, Triple
 
 
 class Loader:
