+++ octadocs/octiron/plugins/markdown.py
@@ -6,7 +6,7 @@ import frontmatter
 import rdflib
 from documented import DocumentedError
 from octadocs.octiron.plugins import Loader
-from octadocs.octiron.types import OCTA, Triple
+from octadocs.types import OCTA, Triple
 from octadocs.octiron.yaml_extensions import as_triple_stream
 from yaml.scanner import ScannerError
