+++ octadocs/octiron/plugins/turtle.py
@@ -3,7 +3,7 @@ from typing import Iterator
 
 import rdflib
 from octadocs.octiron.plugins import Loader
-from octadocs.octiron.types import Triple
+from octadocs.types import Triple
 
 
 class TurtleLoader(Loader):
