+++ tests/test_octiron/test_markdown.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from octadocs.octiron import Octiron
-from octadocs.octiron.types import LOCAL, OCTA
+from octadocs.types import LOCAL, OCTA
 from rdflib import RDF, RDFS, Graph, Literal, URIRef
 
 LOCAL_IRI = URIRef('local:test.md')
